fedimint_dummy_client

Struct DummyClientModule

pub struct DummyClientModule {
    pub(crate) cfg: DummyClientConfig,
    pub(crate) key: Keypair,
    pub(crate) notifier: ModuleNotifier<DummyStateMachine>,
    pub(crate) client_ctx: ClientContext<Self>,
    pub(crate) db: Database,
}

Fields
cfg: DummyClientConfigkey: Keypairnotifier: ModuleNotifier<DummyStateMachine>client_ctx: ClientContext<Self>db: Database
Implementations

impl DummyClientModule

pub async fn print_money_units( &self, amount: Amount, unit: AmountUnit, account_kp: Keypair, ) -> Result<(OperationId, OutPoint)>

pub async fn print_money( &self, amount: Amount, ) -> Result<(OperationId, OutPoint)>

Request the federation prints money for us

pub async fn print_liability( &self, amount: Amount, ) -> Result<(OperationId, OutPoint)>

Use a broken printer to print a liability instead of money If the federation is honest, should always fail

pub async fn send_money( &self, account: PublicKey, amount: Amount, unit: AmountUnit, ) -> Result<OutPoint>

Send money to another user

pub async fn receive_money_hack(&self, outpoint: OutPoint) -> Result<()>

Wait to receive money at an outpoint

pub fn account(&self) -> PublicKey

Return our account

pub async fn get_balance(&self, unit: AmountUnit) -> Result<Amount>

Get balance for a specific amount unit