Trait fedimint_dummy_client::DummyClientExt

pub trait DummyClientExt {
    // Required methods
    fn print_using_account<'life0, 'async_trait>(
        &'life0 self,
        amount: Amount,
        account_kp: KeyPair
    ) -> Pin<Box<dyn Future<Output = Result<(OperationId, OutPoint)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn print_money<'life0, 'async_trait>(
        &'life0 self,
        amount: Amount
    ) -> Pin<Box<dyn Future<Output = Result<(OperationId, OutPoint)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn print_liability<'life0, 'async_trait>(
        &'life0 self,
        amount: Amount
    ) -> Pin<Box<dyn Future<Output = Result<(OperationId, OutPoint)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_money<'life0, 'async_trait>(
        &'life0 self,
        account: XOnlyPublicKey,
        amount: Amount
    ) -> Pin<Box<dyn Future<Output = Result<OutPoint>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn receive_money<'life0, 'async_trait>(
        &'life0 self,
        outpoint: OutPoint
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fed_signature<'life0, 'life1, 'async_trait>(
        &'life0 self,
        message: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Signature>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn account(&self) -> XOnlyPublicKey;
    fn fed_public_key(&self) -> PublicKey;
}

Exposed API calls for client apps

Required Methods

fn print_using_account<'life0, 'async_trait>( &'life0 self, amount: Amount, account_kp: KeyPair ) -> Pin<Box<dyn Future<Output = Result<(OperationId, OutPoint)>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn print_money<'life0, 'async_trait>( &'life0 self, amount: Amount ) -> Pin<Box<dyn Future<Output = Result<(OperationId, OutPoint)>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Request the federation prints money for us

fn print_liability<'life0, 'async_trait>( &'life0 self, amount: Amount ) -> Pin<Box<dyn Future<Output = Result<(OperationId, OutPoint)>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Use a broken printer to print a liability instead of money If the federation is honest, should always fail

fn send_money<'life0, 'async_trait>( &'life0 self, account: XOnlyPublicKey, amount: Amount ) -> Pin<Box<dyn Future<Output = Result<OutPoint>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send money to another user

fn receive_money<'life0, 'async_trait>( &'life0 self, outpoint: OutPoint ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Wait to receive money at an outpoint

fn fed_signature<'life0, 'life1, 'async_trait>( &'life0 self, message: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Signature>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Request the federation signs a message for us

fn account(&self) -> XOnlyPublicKey

Return our account

fn fed_public_key(&self) -> PublicKey

Return the fed’s public key